+++ kifi-backend/modules/search/app/com/keepit/search/SocialGraphInfo.scala
@@ -4,6 +4,7 @@ import com.keepit.search.graph._
 import com.keepit.common.db.Id
 import com.keepit.model.User
 import com.keepit.search.graph.collection.CollectionSearcherWithUser
+import com.keepit.search.graph.bookmark.URIGraphSearcherWithUser
 
 class SocialGraphInfo(userId: Id[User], val uriGraphSearcher: URIGraphSearcherWithUser, val collectionSearcher: CollectionSearcherWithUser, filter: SearchFilter) {
