+++ kifi-backend/modules/search/test/com/keepit/search/MainSearcherTest.scala
@@ -1,7 +1,7 @@
 package com.keepit.search
 
 import com.keepit.scraper.FakeArticleStore
-import com.keepit.search.graph.BookmarkStore
+import com.keepit.search.graph.bookmark._
 import com.keepit.search.index.VolatileIndexDirectoryImpl
 import com.keepit.search.phrasedetector._
 import com.keepit.model._
@@ -19,7 +19,7 @@ import org.apache.lucene.index.IndexWriterConfig
 import com.keepit.search.index.DefaultAnalyzer
 import com.keepit.search.article.ArticleIndexer
 
-import com.keepit.search.graph.{URIGraphImpl, URIGraphIndexer}
+import com.keepit.search.graph.{URIGraphImpl}
 import org.apache.lucene.util.Version
 import com.keepit.search.graph.collection._
 import com.keepit.common.akka.MonitoredAwait
