+++ kifi-backend/modules/search/test/com/keepit/search/graph/GraphTestHelper.scala
@@ -10,6 +10,7 @@ import com.keepit.scraper.FakeArticleStore
 import com.keepit.search.Article
 import com.keepit.search.Lang
 import com.keepit.search.graph.collection._
+import com.keepit.search.graph.bookmark._
 import com.keepit.shoebox.FakeShoeboxServiceClientImpl
 import com.keepit.inject._
 import org.apache.lucene.index.IndexWriterConfig
